Split up ProjectFiles views into List and Tree
