+++ extensions/vscode/webviews/homeView/src/App.vue
@@ -20,7 +20,7 @@ import { ref } from "vue";
 
 import OverlayableView from "src/components/OverlayableView.vue";
 import EvenEasierDeploy from "src/components/EvenEasierDeploy.vue";
-import ProjectFiles from "src/components/views/ProjectFiles.vue";
+import ProjectFiles from "src/components/views/projectFiles/ProjectFiles.vue";
 import PythonPackages from "src/components/views/PythonPackages.vue";
 import RPackages from "src/components/views/RPackages.vue";
 import Credentials from "src/components/views/Credentials.vue";
