+++ extensions/vscode/webviews/homeView/src/components/views/projectFiles/TreeProjectFiles.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>This will be a Tree eventually</div>
+</template>
+
+<script setup lang="ts"></script>
